requirements.txt
Flask
pandas
numpy
scipy
scikit-learn
#zipfile
requests
pillow